Move code to correct place
